Fix Various Post-Release Bugs (#215)
* Fix various bugs

1. set/getFont use StrongRef methods
2. RestoreState(Checked) checks Scissor and BlendState
3. Re-enable soc:u on 3DS
4. Fix renderer corrupting data if bigger than > 4096 verts(?)
5. Add setColorMask nil/none arg
6. Fix up fused games
7. Change logfile.lua to be more useful
8. Use FetchContent_Declare for lua-https
9. Set up cert info in sdmc:/config/ssl/cacert.pem
10. TextShaper fixes

* fix lua-https bit here

* update files for piepie

* Fix up Matrices and Stuff (#216)

* piepie did some things with matrices, fix up checkTextureSettings (with table)

* fix up some more things

* fix this one line in cafe graphics

* should get() the unique_ptr

* why is this still templated

+++ source/objects/font/wrap_font.cpp
@@ -259,7 +259,7 @@ int Wrap_Font::GetWrap(lua_State* L)
         maxWidth = std::max(maxWidth, width);
 
     lua_pushinteger(L, maxWidth);
-    lua_createtable(L, lines.size(), 0);
+    lua_createtable(L, (int)lines.size(), 0);
 
     for (size_t index = 0; index < lines.size(); index++)
     {
